refactor: 첫질문 gpt로 변경 및 로그인 모달 버튼값 조정
